Add zebedee test content
The zebedee test content is all based on content that is publically
accessible via the public website and public APIs, therefore it can be
committed to the repo.

The service token is only used for local testing and was already
commited to the repo as a go constant.

Adding the content directly removes the AWS credentials dependency,
therefore making local environment setup easier. It also makes it easier
to extend the content to add other examples that meet common testing
needs.